feat(coach): archive orchestrate.py, install stub, redirect imports (#531)
