+++ src/atdd/coach/commands/_archived/__init__.py
@@ -1,0 +1,5 @@
+"""Archived decommissioned command modules.
+
+P5 (#531): orchestrate.py archived here for parity-test reuse.
+The live stub at ``commands/orchestrate.py`` prints a migration message.
+"""
