+++ src/atdd/coach/commands/two_phase_commit.py
@@ -39,7 +39,7 @@ from pathlib import Path
 from typing import Optional
 
 from atdd.coach.commands.durability import DecisionWriter
-from atdd.coach.commands.orchestrate import (
+from atdd.coach.commands._archived.orchestrate import (
     PlannedIssue,
     _branch_to_slug,
     _create_worktree,
